Port Dead Letter Channel functionality of Carbon Business Messaging from MB 2.2 to MB 3.0.

+++ components/andes/org.wso2.carbon.andes.admin/src/main/java/org/wso2/carbon/andes/admin/internal/Queue.java
@@ -9,7 +9,7 @@ public class Queue {
 
     private long queueDepth;
 
-    private int messageCount;
+    private long messageCount;
 
     private Calendar createdTime;
 
@@ -40,11 +40,11 @@ public class Queue {
         this.queueDepth = queueDepth;
     }
 
-    public int getMessageCount() {
+    public long getMessageCount() {
         return messageCount;
     }
 
-    public void setMessageCount(int messageCount) {
+    public void setMessageCount(long messageCount) {
         this.messageCount = messageCount;
     }
